refactor ui.lua
